requirements.txt
langchain==0.2.14
langgraph==0.2.28
pydantic==2.8.2
PyYAML==6.0.2
torch>=2.2.0
transformers>=4.43.0
accelerate>=0.33.0
sentence-transformers==3.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==4.9.3
readability-lxml==0.8.1
dateparser==1.2.0
tqdm==4.66.4
tenacity==8.4.2
playwright==1.47.0
pytest==8.3.2
python-json-logger==2.0.7
faiss-cpu==1.12.0